# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.H:
    additionalProperties: true
    type: object
...
  domain.Session:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      ipAddress:
        type: string
      updatedAt:
        type: string
      userAgent:
        type: string
      userId:
        type: string
    type: object
...
  domain.UpdateUserRequest:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      phone:
        type: string
    type: object
  domain.User:
    properties:
      avatarURL:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isEmailVerified:
        type: boolean
      lastName:
        type: string
      password:
        description: Never return password in JSON
        type: string
      phone:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@acareca.com
    name: API Support
  description: API documentation for ACA RECA Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ACA RECA API
  version: "1.0"
paths:
...
  /user-clinic:
    post:
      consumes:
      - application/json
      description: Associate a user with a clinic
      parameters:
      - description: Association request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Associate user with clinic
      tags:
      - UserClinic
  /user-clinic/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a user-clinic association
      parameters:
      - description: User-Clinic Association ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Remove user from clinic
      tags:
      - UserClinic
  /user-clinic/clinic/{clinicId}:
    get:
      consumes:
      - application/json
      description: Get all users associated with a clinic
      parameters:
      - description: Clinic ID
        in: path
        name: clinicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Get clinic's users
      tags:
      - UserClinic
  /user-clinic/user/{userId}:
    get:
      consumes:
      - application/json
      description: Get all clinics associated with a user. Users can only fetch their
        own clinics.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Get user's clinics
      tags:
      - UserClinic
  /users/{userId}:
    get:
      consumes:
      - application/json
      description: Get current user by user ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Get current user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update current user by user ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Update user request
        in: body
        name: updateUserRequest
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Update current user
      tags:
      - User
  /users/{userId}/sessions:
    get:
      consumes:
      - application/json
      description: Get active sessions by user ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Session'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Get active sessions
      tags:
      - User
  /users/{userId}/sessions/{sessionId}:
    delete:
      consumes:
      - application/json
      description: Revoke a session by session ID and user ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Revoke a session
      tags:
      - User
  /users/me:
    get:
      consumes:
      - application/json
      description: Get current authenticated user from JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Get current user (me)
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"